Extract reusable rsjf code to its own file
